add new main.php, add controller index, edit text view index
